requirements.txt
telethon >1.17.0, <1.18.3
future
emoji
requests
sqlalchemy
python-telegram-bot==11.1.0
psycopg2-binary
feedparser
pyowm
bs4
lxml
demjson
PyLyrics
wget
geopy
aiohttp>=2.2.5
gTTS
Pillow>=4.2.0
translations
zalgo_text
nltk
pynewtonmath
bs4
telegraph
psycopg2
tldextract
gsearch
wikipedia
pyowm
PyLyrics
spongemock
psutil
translator
googletrans==3.1.0a0
httpx[http2]
tswift
speedtest-cli
bleach
git+https://github.com/starry69/python-markdown2.git
